requirements.txt
fastapi[all]
uvicorn
openai
python-multipart
pillow
pydantic
httpx
python-dotenv
loguru
jinja2
pydantic-settings
opencv-python-headless
moviepy
numpy
scalar-fastapi
azure-storage-blob
aiohttp